backend/requirements.txt
# Django & REST framework
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.2.2

# Database
psycopg2-binary==2.9.9

# Config management
python-decouple==3.8

# Machine Learning & Image Processing
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
tensorflow==2.16.1
Pillow==10.4.0

# API communication
requests==2.32.3

# For CORS, Auth & Other Django utilities
gunicorn==22.0.0
whitenoise==6.6.0